fix(mcppool): keep pure unit tests in the release gate
Split scope_launcher_test.go: pure argv/env assertions that don't need
systemd now live in scope_launcher_unit_test.go (untagged), so they
remain in the blocking release gate. Only tests that spawn processes or
probe systemd stay behind //go:build integration.

https://github.com/asheshgoplani/agent-deck/pull/1331#issuecomment-4658427463
